+++ src/data/metadata.json
@@ -1,3 +1,3 @@
 {
-    "last_processed_block": 23485569
+    "last_processed_block": 23485764
 }
