+++ src/utils.py
@@ -4,8 +4,12 @@ from dataclasses import asdict
 from os.path import exists
 from sys import maxsize
 
+from algosdk.v2client.indexer import IndexerClient
+
 from .models import (
     AlgoWorldCityAsset,
+    ARC69Attribute,
+    ARC69Record,
     AWENotePrefix,
     StorageMetadata,
     StorageProcessedNote,
@@ -126,3 +130,109 @@ def wait_for_confirmation(client, txid):
         )
     )
     return txinfo
+
+
+def get_onchain_arc(indexer: IndexerClient, address: str, asset_index: int):
+    try:
+        response = indexer.search_transactions(
+            address=address,
+            txn_type="acfg",
+            asset_id=asset_index,
+        )
+
+        if response and "transactions" in response and response["transactions"]:
+            try:
+                asset_config_tx = response["transactions"][0]
+                arc_note = ARC69Record(
+                    **json.loads(
+                        base64.b64decode(asset_config_tx["note"]).decode("utf-8")
+                    )
+                )
+                arc_note.attributes = [
+                    ARC69Attribute(**attribute) for attribute in arc_note.attributes
+                ]
+
+                return arc_note
+
+            except Exception as exp:
+                print(
+                    f"ARC69 not yet configured for city stats for asset index: {asset_index} {exp}"
+                )
+                return None
+    except Exception as exp:
+        print(f"Unable to fetch city stats for {address} {exp}")
+
+    return None
+
+
+def get_onchain_influence(arc_note: ARC69Record):
+
+    if not arc_note:
+        return 0
+
+    for attribute in arc_note.attributes:
+        if attribute.trait_type.lower() == "algoworld influence":
+            return int(attribute.value)
+
+    return -1
+
+
+def get_onchain_city_status(arc_note: ARC69Record):
+
+    if not arc_note:
+        return None
+
+    for attribute in arc_note.attributes:
+        if attribute.trait_type.lower() == "city status":
+            return attribute.value
+
+    return None
+
+
+def get_all_cities(
+    indexer: IndexerClient,
+    manager_address: str,
+    all_assets: list[AlgoWorldCityAsset],
+    awc_prefix: str,
+):
+    all_cities = []
+
+    for asset in all_assets:
+
+        try:
+            print(
+                f'Loading potential city asset under {asset["index"]} and {asset["params"]["name"]}'
+            )
+            asset_index = asset["index"]
+            cur_arc_note = get_onchain_arc(indexer, manager_address, asset_index)
+            cur_influence = get_onchain_influence(cur_arc_note)
+
+            if cur_influence <= 0:
+                print(f"Skipping asset {asset_index} with influence {cur_influence}")
+                continue
+
+            cur_status = get_onchain_city_status(cur_arc_note)
+            cur_status = (
+                get_city_status(cur_influence) if not cur_status else cur_status
+            )
+
+            city = AlgoWorldCityAsset(
+                **{
+                    "index": asset["index"],
+                    "name": asset["params"]["name"],
+                    "url": asset["params"]["url"],
+                    "influence": cur_influence,
+                    "status": cur_status,
+                }
+            )
+            if (
+                len(city.name) > len(awc_prefix)
+                and awc_prefix == city.name[0 : len(awc_prefix)]
+            ):
+                all_cities.append(city)
+            else:
+                print(f"Skipping {city.name} - possibly not an aw city asset")
+        except Exception as exp:
+            print(f"Unable to parse asset: {asset} {exp}. Skipping...")
+
+    return all_cities
